Proofreading lines 200-600. Also used ctrl+F to look for '?!' instead of '!?'
Added an easy point of reference if anyone wants to try translating Paramyres's laugh.
I learned that "a summons" is a noun today, but I wanted to add to line 503 to convey a bit more personality.
Line 592 was worded slightly different when all of the other repeated lines were the same, so I made it the same as the others.